move spatial view into its own crate, separate some utilities into mew space_view crate
