+++ Cargo.toml
@@ -44,6 +44,7 @@ re_renderer = { path = "crates/re_renderer", version = "0.7.0-alpha.0", default-
 re_sdk = { path = "crates/re_sdk", version = "0.7.0-alpha.0", default-features = false }
 re_sdk_comms = { path = "crates/re_sdk_comms", version = "0.7.0-alpha.0", default-features = false }
 re_smart_channel = { path = "crates/re_smart_channel", version = "0.7.0-alpha.0", default-features = false }
+re_space_view = { path = "crates/re_space_view", version = "0.7.0-alpha.0", default-features = false }
 re_space_view_text = { path = "crates/re_space_view_text", version = "0.7.0-alpha.0", default-features = false }
 re_space_view_text_box = { path = "crates/re_space_view_text_box", version = "0.7.0-alpha.0", default-features = false }
 re_space_view_spatial = { path = "crates/re_space_view_spatial", version = "0.7.0-alpha.0", default-features = false }
