+++ crates/re_space_view/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+authors.workspace = true
+description = "Types & utilities for defining Space View classes and communicating with the Viewport."
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "re_space_view"
+publish = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+include = ["../../LICENSE-APACHE", "../../LICENSE-MIT", "**/*.rs", "Cargo.toml"]
+
+[package.metadata.docs.rs]
+all-features = true
+
+[dependencies]
+re_arrow_store.workspace = true
+re_components.workspace = true
+re_data_store.workspace = true
+re_data_ui.workspace = true
+re_log_types.workspace = true
+re_log.workspace = true
+re_query.workspace = true
+re_renderer.workspace = true
+re_tracing.workspace = true
+re_ui.workspace = true
+re_viewer_context.workspace = true
+
+ahash.workspace = true
+egui.workspace = true
+lazy_static.workspace = true
+nohash-hasher.workspace = true
+serde = "1.0"
+slotmap.workspace = true
+smallvec.workspace = true
