+++ crates/re_space_view/README.md
@@ -1,0 +1,10 @@
+# re_space_view
+
+Part of the [`rerun`](https://github.com/rerun-io/rerun) family of crates.
+
+[![Latest version](https://img.shields.io/crates/v/re_space_view.svg)](https://crates.io/crates/re_space_view)
+[![Documentation](https://docs.rs/re_space_view/badge.svg)](https://docs.rs/re_space_view)
+![MIT](https://img.shields.io/badge/license-MIT-blue.svg)
+![Apache](https://img.shields.io/badge/license-Apache-blue.svg)
+
+Types & utilities for defining Space View classes and communicating with the Viewport.
