+++ crates/re_space_view/src/controls.rs
