+++ crates/re_space_view/src/lib.rs
@@ -1,0 +1,12 @@
+//! Rerun Space View utilities
+//!
+//! Types & utilities for defining Space View classes and communicating with the Viewport.
+
+pub mod controls;
+mod data_blueprint;
+mod highlights;
+mod screenshot;
+
+pub use data_blueprint::{DataBlueprintGroup, DataBlueprintTree};
+pub use highlights::{SpaceViewEntityHighlight, SpaceViewHighlights, SpaceViewOutlineMasks};
+pub use screenshot::ScreenshotMode;
