+++ crates/re_space_view_spatial/Cargo.toml
@@ -17,11 +17,38 @@ all-features = true
 
 [dependencies]
 re_arrow_store.workspace = true
-re_components.workspace = true
+re_components = { workspace = true, features = ["ecolor", "glam", "image"] }
+re_data_store = { workspace = true, features = ["serde"] }
+re_data_ui.workspace = true
+re_error.workspace = true
+re_format.workspace = true
+re_log_types.workspace = true
 re_log.workspace = true
 re_query.workspace = true
+re_renderer.workspace = true
+re_string_interner.workspace = true
+re_tracing.workspace = true
 re_ui.workspace = true
 re_viewer_context.workspace = true
+re_space_view.workspace = true
 
+ahash.workspace = true
+anyhow.workspace = true
+bytemuck.workspace = true
+egui-wgpu.workspace = true
 egui.workspace = true
+eframe.workspace = true
+glam.workspace = true
+half.workspace = true
+itertools.workspace = true
+lazy_static.workspace = true
+macaw.workspace = true
+ndarray.workspace = true
+nohash-hasher.workspace = true
+serde = "1"
+slotmap.workspace = true
+smallvec = { workspace = true, features = ["serde"] }
+thiserror.workspace = true
+uuid = { version = "1.1", features = ["serde", "v4", "js"] }
 vec1.workspace = true
+wgpu.workspace = true
