+++ crates/re_space_view_spatial/src/mesh_loader.rs
