+++ crates/re_space_view_spatial/src/scene/scene_part/meshes.rs
@@ -1,13 +1,13 @@
 use re_components::{ColorRGBA, Component, InstanceKey, Mesh3D};
 use re_data_store::EntityPath;
 use re_query::{query_primary_with_history, EntityView, QueryError};
+use re_space_view::{SpaceViewHighlights, SpaceViewOutlineMasks};
 use re_viewer_context::{DefaultColor, SceneQuery, ViewerContext};
 
 use crate::{
     mesh_cache::MeshCache,
-    space_view_highlights::SpaceViewHighlights,
     transform_cache::TransformCache,
-    view_spatial::{scene::EntityDepthOffsets, MeshSource, SceneSpatial},
+    {scene::EntityDepthOffsets, MeshSource, SceneSpatial},
 };
 
 use super::{instance_path_hash_for_picking, ScenePart};
