+++ crates/re_space_view_spatial/src/space_camera_3d.rs
