+++ crates/re_space_view_spatial/src/ui.rs
@@ -8,15 +8,13 @@ use re_data_ui::{item_ui, DataUi};
 use re_data_ui::{show_zoomed_image_region, show_zoomed_image_region_area_outline};
 use re_format::format_f32;
 use re_renderer::OutlineConfig;
+use re_space_view::{DataBlueprintTree, ScreenshotMode, SpaceViewHighlights};
 use re_viewer_context::{
     HoverHighlight, HoveredSpace, Item, SelectionHighlight, SpaceViewId, TensorDecodeCache,
     TensorStatsCache, UiVerbosity, ViewerContext,
 };
 
-use crate::space_view_highlights::SpaceViewHighlights;
-use crate::{
-    data_blueprint::DataBlueprintTree, space_view::ScreenshotMode, view_spatial::UiLabelTarget,
-};
+use crate::UiLabelTarget;
 
 use super::{
     eye::Eye,
