+++ crates/re_space_view_spatial/src/ui_renderer_bridge.rs
