+++ crates/re_viewport/src/lib.rs
@@ -3,30 +3,23 @@
 //! This crate provides the central panel that contains all space views.
 
 mod auto_layout;
-mod data_blueprint;
-mod instance_hash_conversions;
-mod mesh_cache;
-mod mesh_loader;
 mod space_info;
 mod space_view;
-mod space_view_controls;
 mod space_view_entity_picker;
 mod space_view_heuristics;
 mod space_view_highlights;
-mod transform_cache;
 mod view_bar_chart;
 mod view_category;
-mod view_spatial;
 mod view_tensor;
 mod view_time_series;
 mod viewport;
 
 pub mod blueprint_components;
 
+pub use re_space_view_spatial::SpatialNavigationMode; // TODO(andreas): Remove, have two kinds of space views instead.
 pub use space_info::SpaceInfoCollection;
 pub use space_view::{SpaceViewBlueprint, SpaceViewState};
 pub use view_category::ViewCategory;
-pub use view_spatial::SpatialNavigationMode; // TODO(andreas): Remove, have two kinds of space views instead.
 pub use viewport::{Viewport, ViewportState};
 
 // ---------------------------------------------------------------------------
