+++ crates/re_viewport/src/space_info.rs
@@ -1,12 +1,10 @@
-use std::collections::BTreeMap;
-
 use nohash_hasher::IntSet;
+use std::collections::BTreeMap;
 
 use re_arrow_store::{LatestAtQuery, TimeInt, Timeline};
 use re_components::{DisconnectedSpace, Pinhole, Transform3D};
 use re_data_store::{log_db::EntityDb, EntityPath, EntityTree};
-
-use crate::transform_cache::UnreachableTransform;
+use re_space_view_spatial::UnreachableTransform;
 
 /// Transform connecting two space paths.
 #[derive(Clone, Debug)]
