--- crates/re_viewport/src/view_spatial/mod.rs
@@ -1,14 +1,0 @@
-mod eye;
-mod scene;
-mod space_camera_3d;
-
-mod ui;
-pub mod ui_2d;
-pub mod ui_3d;
-pub mod ui_renderer_bridge;
-
-pub use self::scene::{Image, MeshSource, SceneSpatial, UiLabel, UiLabelTarget};
-pub use self::space_camera_3d::SpaceCamera3D;
-pub use ui::{SpatialNavigationMode, ViewSpatialState};
-pub use ui_2d::view_2d;
-pub use ui_3d::{view_3d, SpaceSpecs};
