+++ scripts/publish_crates.sh
@@ -119,6 +119,7 @@ cargo publish $FLAGS -p re_web_viewer_server
 cargo publish $FLAGS -p re_viewer_context
 cargo publish $FLAGS -p re_data_ui
 cargo publish $FLAGS -p re_time_panel
+cargo publish $FLAGS -p re_space_view
 cargo publish $FLAGS -p re_viewport
 cargo publish $FLAGS -p re_space_view_text
 cargo publish $FLAGS -p re_space_view_text_box
